+++ pkg/github/enterprise/module.go
@@ -2,6 +2,7 @@ package enterprise
 
 import (
 	"mash/pkg/di"
+	"mash/pkg/github/enterprise/client"
 	"mash/pkg/github/enterprise/db"
 	"mash/pkg/github/enterprise/service"
 	"mash/pkg/github/enterprise/workers"
@@ -11,6 +12,10 @@ var Module = di.NewModule(
 	di.Needs(db.Module),
 	di.Needs(workers.Module),
 	di.Needs(service.Module),
+
+	di.Provides(func() (client.ClientProvider, client.PersonalClientProvider) {
+		return client.NewClient, client.NewPersonalClient
+	}),
 
 	// todo: hack to solve circular dependency
 	di.Provides(func() *service.ImporterQueue {
