+++ pkg/github/enterprise/routes/push/push.go
@@ -14,10 +14,10 @@ import (
 	"mash/pkg/change/message"
 	workers_ci "mash/pkg/ci/workers"
 	service_comments "mash/pkg/comments/service"
-	"mash/pkg/github/client"
 	"mash/pkg/github/config"
+	"mash/pkg/github/enterprise/client"
 	"mash/pkg/github/enterprise/db"
-	vcs_github "mash/pkg/github/vcs"
+	vcs_github "mash/pkg/github/enterprise/vcs"
 	db_review "mash/pkg/review/db"
 	service_sync "mash/pkg/sync/service"
 	"mash/pkg/view/events"
