+++ pkg/github/enterprise/routes/webhook.go
@@ -8,8 +8,8 @@ import (
 	workers_ci "mash/pkg/ci/workers"
 	db_codebase "mash/pkg/codebase/db"
 	service_comments "mash/pkg/comments/service"
-	"mash/pkg/github/client"
 	"mash/pkg/github/config"
+	"mash/pkg/github/enterprise/client"
 	"mash/pkg/github/enterprise/db"
 	"mash/pkg/github/enterprise/routes/installation"
 	"mash/pkg/github/enterprise/routes/pr"
