+++ pkg/github/enterprise/service/service.go
@@ -3,21 +3,21 @@ package service
 import (
 	"context"
 	"fmt"
+	"time"
+
+	"mash/pkg/change"
 	db_codebase "mash/pkg/codebase/db"
+	"mash/pkg/github"
+	config_github "mash/pkg/github/config"
+	github_client "mash/pkg/github/enterprise/client"
+	db_github "mash/pkg/github/enterprise/db"
+	github_vcs "mash/pkg/github/enterprise/vcs"
 	"mash/pkg/notification/sender"
 	"mash/pkg/snapshots/snapshotter"
 	service_user "mash/pkg/user/service"
 	"mash/pkg/view/events"
 	db_workspace "mash/pkg/workspace/db"
 	"mash/vcs"
-	"time"
-
-	"mash/pkg/change"
-	"mash/pkg/github"
-	github_client "mash/pkg/github/client"
-	config_github "mash/pkg/github/config"
-	db_github "mash/pkg/github/enterprise/db"
-	github_vcs "mash/pkg/github/vcs"
 	"mash/vcs/executor"
 
 	"github.com/posthog/posthog-go"
