+++ pkg/github/enterprise/service/service_import.go
@@ -5,17 +5,18 @@ import (
 	"database/sql"
 	"errors"
 	"fmt"
+	"time"
+
 	"mash/pkg/change/message"
 	"mash/pkg/github"
-	github_client "mash/pkg/github/client"
-	github_vcs "mash/pkg/github/vcs"
+	github_client "mash/pkg/github/enterprise/client"
+	github_vcs "mash/pkg/github/enterprise/vcs"
 	"mash/pkg/snapshots"
 	"mash/pkg/snapshots/snapshotter"
 	view_vcs "mash/pkg/view/vcs"
 	"mash/pkg/workspace"
 	"mash/vcs"
 	"mash/vcs/provider"
-	"time"
 
 	gh "github.com/google/go-github/v39/github"
 	"github.com/google/uuid"
