+++ pkg/github/graphql/enterprise/codebase_github_integration.go
@@ -6,8 +6,8 @@ import (
 	service_auth "mash/pkg/auth/service"
 	service_codebase "mash/pkg/codebase/service"
 	"mash/pkg/github"
-	github_client "mash/pkg/github/client"
 	"mash/pkg/github/config"
+	github_client "mash/pkg/github/enterprise/client"
 	db_github "mash/pkg/github/enterprise/db"
 	service_github "mash/pkg/github/enterprise/service"
 	gqlerrors "mash/pkg/graphql/errors"
