+++ pkg/github/graphql/pr/enterprise/pr_root_resolver.go
@@ -8,8 +8,8 @@ import (
 	"mash/pkg/auth"
 	service_auth "mash/pkg/auth/service"
 	db_codebase "mash/pkg/codebase/db"
-	"mash/pkg/github/client"
 	"mash/pkg/github/config"
+	"mash/pkg/github/enterprise/client"
 	"mash/pkg/github/enterprise/db"
 	service_github "mash/pkg/github/enterprise/service"
 	gqlerrors "mash/pkg/graphql/errors"
