+++ pkg/github/graphql/pr/enterprise/pr_test.go
@@ -29,8 +29,8 @@ import (
 	graphql_comments "mash/pkg/comments/graphql"
 	service_comments "mash/pkg/comments/service"
 	"mash/pkg/github"
-	"mash/pkg/github/client"
 	"mash/pkg/github/config"
+	"mash/pkg/github/enterprise/client"
 	db_github "mash/pkg/github/enterprise/db"
 	"mash/pkg/github/enterprise/routes"
 	service_github "mash/pkg/github/enterprise/service"
