+++ pkg/http/enterprise/handler.go
@@ -8,8 +8,8 @@ import (
 	workers_ci "mash/pkg/ci/workers"
 	db_codebase "mash/pkg/codebase/db"
 	service_comments "mash/pkg/comments/service"
-	ghappclient "mash/pkg/github/client"
 	"mash/pkg/github/config"
+	ghappclient "mash/pkg/github/enterprise/client"
 	db_github "mash/pkg/github/enterprise/db"
 	routes_v3_ghapp "mash/pkg/github/enterprise/routes"
 	service_github "mash/pkg/github/enterprise/service"
